+++ lux-cli/src/lib.rs
@@ -43,7 +43,6 @@ pub mod check;
 pub mod completion;
 pub mod config;
 pub mod debug;
-pub mod dependencies;
 pub mod dist;
 pub mod doc;
 pub mod download;
@@ -70,6 +69,7 @@ pub mod search;
 pub mod shell;
 pub mod sync;
 pub mod test;
+pub mod toolchains;
 pub mod uninstall;
 pub mod unpack;
 pub mod update;
